Use Double everywhere instead of Float.

+++ src/main/scala/org/jesperdj/scalaray/Main.scala
@@ -50,7 +50,7 @@ object Main {
 
     val rect = new Rectangle(0, 0, 799, 599)
 
-    val camera: Camera = new PerspectiveCamera(Transform.translate(0.0f, 0.75f, 0.0f), π / 4.0f, rect.width, rect.height)
+    val camera: Camera = new PerspectiveCamera(Transform.translate(0.0, 0.75f, 0.0), π / 4.0, rect.width, rect.height)
 
     val filter: Filter = new BoxFilter
     val raster = new PixelRaster(rect, filter)
@@ -82,34 +82,34 @@ object Main {
   }
 
   def createScene(): Scene = {
-    val mat1 = new MatteMaterial(new ConstantTexture(new Spectrum(1.0f, 1.0f, 0.5f)))
-    val mat2 = new MatteMaterial(new ConstantTexture(new Spectrum(1.0f, 1.0f, 1.0f)))
+    val mat1 = new MatteMaterial(new ConstantTexture(new Spectrum(1.0, 1.0, 0.5)))
+    val mat2 = new MatteMaterial(new ConstantTexture(new Spectrum(1.0, 1.0, 1.0)))
 
-//    val s1 = new Sphere(0.75f, -1.0f, 1.0f, π * 4.0f / 3.0f)
-//    val p1 = new TransformedPrimitive(new GeometricPrimitive(s1, mat1), Transform.translate(0.0f, 0.75, 4.0f) * Transform.rotateY(π / 4.0f) * Transform.rotateZ(π / 2.0f))
-    val p1 = new TransformedPrimitive(createCube(0.5f, mat1), Transform.translate(0.0f, 0.75f, 4.0f) * Transform.rotateY(-π / 6.0f) * Transform.rotateX(-π / 6.0f))
+//    val s1 = new Sphere(0.75f, -1.0, 1.0, π * 4.0 / 3.0)
+//    val p1 = new TransformedPrimitive(new GeometricPrimitive(s1, mat1), Transform.translate(0.0, 0.75, 4.0) * Transform.rotateY(π / 4.0) * Transform.rotateZ(π / 2.0))
+    val p1 = new TransformedPrimitive(createCube(0.5, mat1), Transform.translate(0.0, 0.75f, 4.0) * Transform.rotateY(-π / 6.0) * Transform.rotateX(-π / 6.0))
 
-    val s2 = new Disk(3.0f)
-    val p2 = new TransformedPrimitive(new GeometricPrimitive(s2, mat2), Transform.translate(0.0f, 0.0f, 4.0f) * Transform.rotateX(-π / 2.0f))
+    val s2 = new Disk(3.0)
+    val p2 = new TransformedPrimitive(new GeometricPrimitive(s2, mat2), Transform.translate(0.0, 0.0, 4.0) * Transform.rotateX(-π / 2.0))
 
-    val l1 = new PointLightSource(new Point(0.5f, 2.0f, 0.0f), new Spectrum(30.0f, 30.0f, 30.0f))
+    val l1 = new PointLightSource(new Point(0.5, 2.0, 0.0), new Spectrum(30.0, 30.0, 30.0))
 
-    val s3 = new Disk(1.5f)
-    val t3 = Transform.translate(-0.3f, 5.0f, 3.5f) * Transform.rotateX(π / 2.0f)
+    val s3 = new Disk(1.5)
+    val t3 = Transform.translate(-0.3f, 5.0, 3.5) * Transform.rotateX(π / 2.0)
     val l2 = new DiffuseAreaLightSource(s3, t3, new Spectrum(0.7f, 0.7f, 0.7f), 4)
     val p3 = new TransformedPrimitive(new GeometricPrimitive(s3, mat1, l2), t3)
 
     new Scene(new CompositePrimitive(p1, p2, p3), Traversable(l1, l2))
   }
 
-  def createCube(size: Float, material: Material): Primitive = {
+  def createCube(size: Double, material: Material): Primitive = {
     import scala.collection.immutable.IndexedSeq
 
     // TODO: Properly generate normals and u, v surface coordinates for vertices
 
     // Generate vertices
     val v: IndexedSeq[Vertex] = for (i <- 0 to 7) yield
-      new Vertex(Point(if ((i & 1) != 0) size else -size, if ((i & 2) != 0) size else -size, if ((i & 4) != 0) size else -size), Normal.Zero, 0.0f, 0.0f)
+      new Vertex(Point(if ((i & 1) != 0) size else -size, if ((i & 2) != 0) size else -size, if ((i & 4) != 0) size else -size), Normal.Zero, 0.0, 0.0)
 
     // Front
     val ft1 = new GeometricPrimitive(new Triangle(v(0), v(2), v(3)), material)
