+++ src/main/scala/org/jesperdj/scalaray/scene/BoundingVolumeHierarchyAccelerator.scala
@@ -47,7 +47,7 @@ final class BoundingVolumeHierarchyAccelerator (
   override def boundingBox(transform: Transform): BoundingBox = root.boundingBox(transform)
 
   // Compute closest intersection between a ray and this primitive, returns intersection and and distance of intersection along ray
-  def intersect(ray: Ray): Option[(Intersection, Float)] = root intersect ray
+  def intersect(ray: Ray): Option[(Intersection, Double)] = root intersect ray
 
   // Check if a ray intersects this primitive
   override def checkIntersect(ray: Ray): Boolean = root checkIntersect ray
